requirements.txt
faiss_cpu==1.7.4
langchain==0.0.334
numpy==1.24.2
openai==0.28.1
openai[embeddings]==0.28.1
python-dotenv==1.0.0
tqdm==4.65.0
tiktoken==0.4.0
pandas==2.0.3
scikit-learn==1.2.2
plotly==5.15.0
umap-learn==0.5.3
sentence-transformers==2.2.2
textwrap3==0.9.2
krippendorff==0.6.0
beautifulsoup4==4.12.2
func_timeout==4.3.5
transformers==4.34.1
cohere==4.34
bitsandbytes==0.41.2
accelerate==0.24.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
seaborn==0.13.0
vllm==0.2.0
huggingface_hub==0.17.3
googlesearch-python==1.2.3
wikipedia==1.4.0